Bug 1741196 - Modify MemoryDistribution::Accumulate() to take size_t instead of u_int64 r=chutten
Differential Revision: https://phabricator.services.mozilla.com/D152308

UltraBlame original commit: 2e6f8694537a4c2a6d35ec4931e6de5214254687

+++ toolkit/components/glean/bindings/private/MemoryDistribution.h
@@ -29,7 +29,7 @@ class MemoryDistributionMetric {
 
 
 
-  void Accumulate(uint64_t aSample) const;
+  void Accumulate(size_t aSample) const;
